requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==5.1.4
django-celery-beat==2.7.0
django-import-export==4.3.3
django-timezone-field==7.0
et_xmlfile==2.0.0
idna==3.10
kombu==5.4.2
openpyxl==3.1.5
prompt_toolkit==3.0.48
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
